Pipeline: refine directory for pipeline (#9100)
ref pingcap/tiflash#6233
