+++ dbms/src/Flash/CMakeLists.txt
@@ -22,8 +22,10 @@ add_headers_and_sources(flash_service ./Pipeline)
 add_headers_and_sources(flash_service ./Pipeline/Exec)
 add_headers_and_sources(flash_service ./Pipeline/Schedule)
 add_headers_and_sources(flash_service ./Pipeline/Schedule/Events)
+add_headers_and_sources(flash_service ./Pipeline/Schedule/Events/Impls)
 add_headers_and_sources(flash_service ./Pipeline/Schedule/Reactor)
 add_headers_and_sources(flash_service ./Pipeline/Schedule/Tasks)
+add_headers_and_sources(flash_service ./Pipeline/Schedule/Tasks/Impls)
 add_headers_and_sources(flash_service ./Pipeline/Schedule/TaskQueues)
 add_headers_and_sources(flash_service ./Pipeline/Schedule/ThreadPool)
 add_headers_and_sources(flash_service ./Planner)
