+++ dbms/src/Flash/Pipeline/Schedule/Tasks/Impls/EventTask.cpp
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include <Common/Exception.h>
-#include <Flash/Pipeline/Schedule/Tasks/EventTask.h>
+#include <Flash/Pipeline/Schedule/Tasks/Impls/EventTask.h>
 
 namespace DB
 {
