+++ dbms/src/Flash/Pipeline/Schedule/Tasks/Impls/IOEventTask.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <Flash/Pipeline/Schedule/Tasks/EventTask.h>
+#include <Flash/Pipeline/Schedule/Tasks/Impls/EventTask.h>
 
 namespace DB
 {
