+++ dbms/src/Flash/Pipeline/Schedule/Tasks/Impls/PipelineTask.h
@@ -15,8 +15,8 @@
 #pragma once
 
 #include <Flash/Pipeline/Exec/PipelineExec.h>
-#include <Flash/Pipeline/Schedule/Tasks/EventTask.h>
-#include <Flash/Pipeline/Schedule/Tasks/PipelineTaskBase.h>
+#include <Flash/Pipeline/Schedule/Tasks/Impls/EventTask.h>
+#include <Flash/Pipeline/Schedule/Tasks/Impls/PipelineTaskBase.h>
 
 namespace DB
 {
