+++ dbms/src/Flash/Pipeline/Schedule/Tasks/Impls/SimplePipelineTask.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include <Flash/Pipeline/Exec/PipelineExec.h>
-#include <Flash/Pipeline/Schedule/Tasks/PipelineTaskBase.h>
+#include <Flash/Pipeline/Schedule/Tasks/Impls/PipelineTaskBase.h>
 #include <Flash/Pipeline/Schedule/Tasks/Task.h>
 
 namespace DB
